+++ _inc/blocks/business-hours/view.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array('wp-polyfill'), 'version' => '8555d430584123112017a11e9e5a5d41');
+<?php return array('dependencies' => array('wp-polyfill'), 'version' => 'a13a4d89fd545ca3d637fbec755da573');