+++ _inc/blocks/dialogue/view.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array('wp-data', 'wp-dom-ready', 'wp-polyfill'), 'version' => '8bdb6bde530d81ceafe2c7b6d2161074');
+<?php return array('dependencies' => array('wp-data', 'wp-dom-ready', 'wp-polyfill'), 'version' => '8d2ec19952e1fed966c56ab149d1505a');