+++ _inc/blocks/image-compare/view.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array('wp-dom-ready', 'wp-polyfill'), 'version' => '42dfe0eb4225098038ab3138670ddcee');
+<?php return array('dependencies' => array('wp-dom-ready', 'wp-polyfill'), 'version' => '1a69b5215b2dfbf7cbc9313deda71c0e');