+++ _inc/blocks/slideshow/view.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array('lodash', 'wp-dom-ready', 'wp-escape-html', 'wp-polyfill'), 'version' => '9e0140f6dbf698741b5e42d727fb8a7b');
+<?php return array('dependencies' => array('lodash', 'wp-dom-ready', 'wp-escape-html', 'wp-polyfill'), 'version' => '0eda3da309e21d5e1e8e37e7cfc63f97');